Add mixin for register dialog, shared logic
