+++ client/components/Main/Tournament/Active/index.vue
@@ -3,7 +3,7 @@
     <tournament-card
       v-for="game in games"
       :key="game.id"
-      @register-dialog="registerDialog"
+      @register-dialog="registerDialog(game.id)"
       @show-details="showDetails"
       :title-list="titleList"
       :game="game"
@@ -22,6 +22,8 @@ import { getGameCarousel, getRanking } from '@/helpers/api/UserGames';
 import { gameExists } from '@/helpers/mixins/BooleanMixin';
 import { mapGetters } from 'vuex';
 import RankingModal from '../ranking/RankingForm';
+import { registerDialog } from '@/helpers/mixins/RegisterDialog';
+import { showModal } from '@/helpers/mixins/RankingModal';
 import TournamentCard from '../card/TournamentCard';
 import TournamentMessage from '../message/TournamentErrorMessage';
 
@@ -40,35 +42,9 @@ const detailTitleList = [
 
 export default {
   name: 'active-tournament',
-  mixins: [gameExists],
+  mixins: [gameExists, registerDialog, showModal],
   computed: { ...mapGetters({ games: 'activeGameData' }) },
   methods: {
-    submitIdHandler(id) {
-      const submittedId = Math.floor((Math.random() * 10000) + 1);
-      this.$store.dispatch('submitPlayerId',
-        { activeGameId: id, playerId: submittedId });
-    },
-    registerDialog(id) {
-      this.$modal.show('dialog', {
-        title: 'Please confirm your registration',
-        text: 'To enter the tournament, you have to confirm your application.',
-        buttons: [
-          {
-            title: 'Confirm',
-            handler: () => {
-              this.submitIdHandler(id);
-              this.$modal.hide('dialog');
-            }
-          },
-          {
-            title: 'Cancel',
-            handler: () => {
-              this.$modal.hide('dialog');
-            }
-          }
-        ]
-      });
-    },
     showDetails(data) {
       const {
         playerId, id, image, title, rules,
